+++ dashboard/modules/metrics/dashboards/serve_grafana_dashboard_base.json
@@ -52,7 +52,7 @@
           ]
         },
         "datasource": "${datasource}",
-        "definition": "label_values({{__name__=~ 'ray_serve_num_http_requests|ray_serve_num_grpc_requests',{global_filters}}}, application)",
+        "definition": "label_values(ray_serve_deployment_replica_healthy{{{global_filters}}}, application)",
         "description": null,
         "error": null,
         "hide": 0,
@@ -62,7 +62,7 @@
         "name": "Application",
         "options": [],
         "query": {
-          "query": "label_values({{__name__=~ 'ray_serve_num_http_requests|ray_serve_num_grpc_requests',{global_filters}}}, application)",
+          "query": "label_values(ray_serve_deployment_replica_healthy{{{global_filters}}}, application)",
           "refId": "Prometheus-Instance-Variable-Query"
         },
         "refresh": 2,
