+++ python/ray/serve/tests/test_cli_2.py
@@ -28,7 +28,7 @@ from ray.serve._private.constants import (
 )
 from ray.serve.generated import serve_pb2, serve_pb2_grpc
 import grpc
-from ray.serve.tests.test_grpc import (
+from ray.serve.tests.utils import (
     ping_grpc_list_applications,
     ping_grpc_healthz,
     ping_grpc_call_method,
